add function save NhanVienModel

+++ src/com/polymart/dao/impl/AbstractDAO.java
@@ -117,12 +117,12 @@ public class AbstractDAO<T> implements GenericDAO<T> {
     }
 
     @Override
-    public Long insert(String sql, Object... parameters) {
+    public Integer insert(String sql, Object... parameters) {
         Connection connection = null;
         PreparedStatement statement = null;
         ResultSet resultSet = null;
         try {
-            Long id = null;
+            Integer id = null;
             connection = getConnection();
             connection.setAutoCommit(false);
             statement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS);
@@ -130,7 +130,7 @@ public class AbstractDAO<T> implements GenericDAO<T> {
             statement.executeUpdate();
             resultSet = statement.getGeneratedKeys();
             if (resultSet.next()) {
-                id = resultSet.getLong(1);
+                id = resultSet.getInt(1);
             }
             connection.commit();
             return id;
